+++ client/components/Form.css
@@ -1,0 +1,14 @@
+.fade-enter {
+  transform: translateX(100%);
+}
+.fade-enter-active {
+  transform: translateX(0);
+  transition: transform 300ms;
+}
+.fade-exit {
+  transform: translateX(0);
+}
+.fade-exit-active {
+  transform: translateX(-100%);
+  transition: transform 300ms;
+}
